feat: migrate data and domain to WorldMap.Common
