--- src/RepositoryBase/RepositoryBase.sln
@@ -1,41 +1,0 @@
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio 15
-VisualStudioVersion = 15.0.26730.12
-MinimumVisualStudioVersion = 15.0.26124.0
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{34DE985D-D09B-4D8C-89C5-5B94332CA014}"
-EndProject
-Project("{9A19103F-16F7-4668-BE54-9A1E7A4F7556}") = "RepositoryBase", "src\RepositoryBase\RepositoryBase.csproj", "{F47E5D9A-1798-4154-8BCA-535A435DCB5C}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Debug|x64 = Debug|x64
-		Debug|x86 = Debug|x86
-		Release|Any CPU = Release|Any CPU
-		Release|x64 = Release|x64
-		Release|x86 = Release|x86
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Debug|x64.ActiveCfg = Debug|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Debug|x64.Build.0 = Debug|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Debug|x86.ActiveCfg = Debug|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Debug|x86.Build.0 = Debug|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Release|x64.ActiveCfg = Release|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Release|x64.Build.0 = Release|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Release|x86.ActiveCfg = Release|Any CPU
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C}.Release|x86.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{F47E5D9A-1798-4154-8BCA-535A435DCB5C} = {34DE985D-D09B-4D8C-89C5-5B94332CA014}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {C1BC7DF8-24A1-4932-9EFF-67C00ECBD74C}
-	EndGlobalSection
-EndGlobal
