--- src/World.Data.Xml/World.Data.Xml.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFramework>netstandard1.3</TargetFramework>
-    <Description>World countries data repositories for XML data source.</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\World.Data.Common\World.Data.Common.csproj" />
-    <ProjectReference Include="..\World.Data.Models\World.Data.Models.csproj" />
-  </ItemGroup>
-
-</Project>
