--- src/World.Data/World.Data.csproj
@@ -1,13 +1,0 @@
-<Project Sdk="Microsoft.NET.Sdk">
-
-  <PropertyGroup>
-    <TargetFrameworks>netstandard2.0</TargetFrameworks>
-    <Description>World countries data repositories for Sql data source.</Description>
-  </PropertyGroup>
-
-  <ItemGroup>
-    <ProjectReference Include="..\World.Data.Common\World.Data.Common.csproj" />
-    <ProjectReference Include="..\World.Data.Models\World.Data.Models.csproj" />
-  </ItemGroup>
-  
-</Project>
