+++ src/WorldMap.Common/Data/IRepositoryGetByKey.cs
@@ -1,0 +1,7 @@
+﻿namespace WorldMap.Common.Data
+{
+	public interface IRepositoryGetByKey<TKey, T> where T : class
+	{
+		T Get(TKey id);
+	}
+}
